level is float in db

+++ bot/src/db/models.go
@@ -30,7 +30,7 @@ type level struct {
 	User_ID			string	`bun:"user_id,notnull"`
 	Guild_ID		string	`bun:"guild_id,notnull"`
 	Nb_Msg			int64	`bun:"nb_msg,notnull,default:1"`
-	Level			int64	`bun:"level,notnull,default:0"`
+	Level			float64	`bun:"level,notnull,default:0"`
 }
 
 type action struct {
